requirements.txt
openai
langchain
langchain-community
langchain-text-splitters
langchain-openai
chromadb
pypdf
python-dotenv
fastapi
uvicorn
tiktoken